Refining regex
Fixed... pp was being removed from words.
Fixed... any music notes were being changed to [music] [music] even
if there was only one note. We are switching back to only looking
for a group of notes or a single one. If there are two music notes
with a space between them, then it will appear as [music] [music].
We will just have to deal with it and let anyone captioning know that
they should only use one music not... or preferrably tell them just to
type [music].